Cleanup MissionData and move class into Animate3DSymbols files
